refactor: extract tool definitions into tools/ subpackage
Move all 16 MCP tool definitions from __main__.py (~956 lines) into a
dedicated tools/ subpackage. Each tool category gets its own module:

  tools/__init__.py          - register_all() orchestrator
  tools/_snapshot_helpers.py - shared capture/response helpers
  tools/app.py               - App (launch/resize/switch)
  tools/shell.py             - PowerShell
  tools/filesystem.py        - FileSystem (8 modes)
  tools/snapshot.py          - Snapshot + Screenshot
  tools/input.py             - Click, Type, Scroll, Move, Shortcut, Wait
  tools/scrape.py            - Scrape (async, with LLM sampling)
  tools/multi.py             - MultiSelect, MultiEdit
  tools/clipboard.py         - Clipboard
  tools/process.py           - Process
  tools/notification.py      - Notification
  tools/registry.py          - Registry

Architecture:
- Each module exports register(mcp, *, get_desktop, get_analytics)
- get_desktop / get_analytics are zero-arg callables for lazy resolution
- __main__.py becomes a thin entrypoint (~140 lines): imports, globals,
  lifespan, mcp creation, register_all() call, CLI
- Backward-compatible re-exports (state_tool, screenshot_tool) preserved
  for existing test_snapshot_display_filter.py

All existing tests pass (no regressions introduced).
