+++ src/windows_mcp/tools/shell.py
@@ -1,0 +1,26 @@
+"""PowerShell tool — shell/command execution."""
+
+from mcp.types import ToolAnnotations
+from windows_mcp.analytics import with_analytics
+from fastmcp import Context
+
+
+def register(mcp, *, get_desktop, get_analytics):
+    @mcp.tool(
+        name="PowerShell",
+        description="Shell/command execution. Keywords: shell, run, execute, cmd, terminal, command line, script. A comprehensive system tool for executing any PowerShell commands. Use it to navigate the file system, manage files and processes, and execute system-level operations. Capable of accessing web content (e.g., via Invoke-WebRequest), interacting with network resources, and performing complex administrative tasks. This tool provides full access to the underlying operating system capabilities, making it the primary interface for system automation, scripting, and deep system interaction.",
+        annotations=ToolAnnotations(
+            title="PowerShell",
+            readOnlyHint=False,
+            destructiveHint=True,
+            idempotentHint=False,
+            openWorldHint=True,
+        ),
+    )
+    @with_analytics(get_analytics(), "Powershell-Tool")
+    def powershell_tool(command: str, timeout: int = 30, ctx: Context = None) -> str:
+        try:
+            response, status_code = get_desktop().execute_command(command, timeout)
+            return f"Response: {response}\nStatus Code: {status_code}"
+        except Exception as e:
+            return f"Error executing command: {str(e)}\nStatus Code: 1"
